Refactor examples: common simple scene controller
